exercise 4.21

+++ part4/bloglist/controllers/blogs.js
@@ -39,6 +39,16 @@ blogsRouter.put("/:id", async (request, response) => {
 });
 
 blogsRouter.delete("/:id", async (request, response) => {
+  const decodedToken = jwt.verify(request.token, process.env.SECRET);
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: "token invalid" });
+  }
+  const user = await User.findById(decodedToken.id);
+  const blogToDelete = await Blog.findById(request.params.id);
+
+  if (blogToDelete.user.toString() !== user.id) {
+    return response.status(401).json({ error: "unauthorized" });
+  }
   await Blog.findByIdAndRemove(request.params.id);
   response.status(204).end();
 });
